internal api + test_Encryption.cc
