+++ src/librbd/api/Image.h
@@ -72,7 +72,7 @@ struct Image {
 
   static int encryption_format(ImageCtxT *ictx, encryption_format_t format,
                                encryption_options_t opts, size_t opts_size,
-                               bool c_api);
+                               bool insecure_fast_mode, bool c_api);
   static int encryption_load(ImageCtxT *ictx, const encryption_spec_t *specs,
                              size_t spec_count, bool c_api);
 };
