+++ src/librbd/crypto/EncryptionFormat.h
@@ -19,7 +19,8 @@ struct EncryptionFormat {
   }
 
   virtual std::unique_ptr<EncryptionFormat<ImageCtxT>> clone() const = 0;
-  virtual void format(ImageCtxT* ictx, Context* on_finish) = 0;
+  virtual void format(ImageCtxT* ictx, bool insecure_fast_mode,
+                      Context* on_finish) = 0;
   virtual void load(ImageCtxT* ictx, std::string* detected_format_name,
                     Context* on_finish) = 0;
   virtual void flatten(ImageCtxT* ictx, Context* on_finish) = 0;
