+++ src/librbd/crypto/FormatRequest.h
@@ -21,11 +21,14 @@ public:
     using EncryptionFormat = decltype(I::encryption_format);
 
     static FormatRequest* create(
-            I* image_ctx, EncryptionFormat format, Context* on_finish) {
-      return new FormatRequest(image_ctx, std::move(format), on_finish);
+            I* image_ctx, EncryptionFormat format, bool insecure_fast_mode,
+            Context* on_finish) {
+      return new FormatRequest(image_ctx, std::move(format),
+                               insecure_fast_mode, on_finish);
     }
 
-    FormatRequest(I* image_ctx, EncryptionFormat format, Context* on_finish);
+    FormatRequest(I* image_ctx, EncryptionFormat format,
+                  bool insecure_fast_mode, Context* on_finish);
     void send();
     void handle_shutdown_crypto(int r);
     void format();
@@ -38,6 +41,7 @@ private:
     I* m_image_ctx;
 
     EncryptionFormat m_format;
+    bool m_insecure_fast_mode;
     Context* m_on_finish;
 };
