+++ src/test/librbd/CMakeLists.txt
@@ -12,6 +12,7 @@ set(librbd_test
   test_internal.cc
   test_mirroring.cc
   test_DeepCopy.cc
+  test_Encryption.cc
   test_Groups.cc
   test_Migration.cc
   test_MirroringWatcher.cc
