+++ src/test/librbd/mock/crypto/MockEncryptionFormat.h
@@ -14,7 +14,7 @@ namespace crypto {
 
 struct MockEncryptionFormat {
   MOCK_CONST_METHOD0(clone, std::unique_ptr<MockEncryptionFormat>());
-  MOCK_METHOD2(format, void(MockImageCtx*, Context*));
+  MOCK_METHOD3(format, void(MockImageCtx*, bool, Context*));
   MOCK_METHOD3(load, void(MockImageCtx*, std::string*, Context*));
   MOCK_METHOD2(flatten, void(MockImageCtx*, Context*));
   MOCK_METHOD0(get_crypto, MockCryptoInterface*());
