+++ src/test/librbd/test_fixture.h
@@ -30,8 +30,8 @@ public:
   int snap_create(librbd::ImageCtx &ictx, const std::string &snap_name);
   int snap_protect(librbd::ImageCtx &ictx, const std::string &snap_name);
 
-  int flatten(librbd::ImageCtx &ictx, librbd::ProgressContext &prog_ctx);
-  int resize(librbd::ImageCtx *ictx, uint64_t size);
+  static int flatten(librbd::ImageCtx *ictx);
+  static int resize(librbd::ImageCtx *ictx, uint64_t size);
 
   int lock_image(librbd::ImageCtx &ictx, ClsLockType lock_type,
                  const std::string &cookie);
